+++ src/components/common/Banner.js
@@ -1,11 +1,10 @@
 import React, { useState, useEffect } from "react";
-import styled, { css } from "styled-components";
-import { ToastContainer, toast } from "react-toastify";
-import { FaChevronLeft, FaChevronRight } from "react-icons/fa";
+import styled from "styled-components";
 
 import { jobs_hero, nearby_stores } from "../../assets/bannerImages";
+import Carousel from "./Carousel";
 
-const Carousel = () => {
+const Banner = () => {
   const example = [
     {
       title: ["남에게 더 가치있게.", <br />, "만원의 행복  🛍️"],
@@ -42,82 +41,20 @@ const Carousel = () => {
     };
   }, []);
 
-  const onIncreaseIdx = () => {
-    if (currItemIndex >= example.length - 1) {
-      setCurrItemIndex(0);
-    } else {
-      setCurrItemIndex(currItemIndex + 1);
-    }
-  };
-  const onDecreaseIdx = () => {
-    if (currItemIndex <= 0) {
-      setCurrItemIndex(example.length - 1);
-    } else {
-      setCurrItemIndex(currItemIndex - 1);
-    }
-  };
-
-  // useEffect(() => {
-  //   const interval = setInterval(() => {
-  //     onIncreaseIdx();
-  //     console.log(currItemIndex);
-  //   }, 2000);
-
-  //   return () => {
-  //     clearInterval(interval);
-  //   };
-  // }, []);
   return (
     <>
       <BannerBlock>
-        <div className="carousel">
-          <div
-            className="banner-container"
-            style={{ transform: `translateX(-${currItemIndex * 100}%)` }}
-          >
-            {example.map((banner, idx) => (
-              <div key={idx} className="banner-wrapper">
-                <span className="banner-text">
-                  <strong className="title">{banner.title}</strong>
-                  <p className="content">{banner.content}</p>
-                </span>
-                <span className="banner-img">
-                  <img src={banner.img} alt="banner" />
-                </span>
-              </div>
-            ))}
-          </div>
-          <p className="icons left">
-            <FaChevronLeft onClick={onDecreaseIdx} />
-          </p>
-          <p className="icons right">
-            <FaChevronRight onClick={onIncreaseIdx} />
-          </p>
-        </div>
-
-        <CarouselDots
+        <Carousel
+          example={example}
           currItemIndex={currItemIndex}
-          itemLength={example.length}
+          setCurrItemIndex={setCurrItemIndex}
         />
       </BannerBlock>
-      {/* <ToastContainer position={toast.POSITION.TOP_CENTER} /> */}
     </>
   );
 };
 
-const CarouselDots = ({ currItemIndex, itemLength }) => {
-  return (
-    <DotsBlock>
-      {Array(itemLength)
-        .fill(null)
-        .map((_, idx) => (
-          <Dot key={idx} active={currItemIndex === idx} />
-        ))}
-    </DotsBlock>
-  );
-};
-
-export default Carousel;
+export default Banner;
 
 const BannerBlock = styled.div`
   display: flex;
@@ -142,17 +79,17 @@ const BannerBlock = styled.div`
       position: absolute; // 아이콘들을 절대 위치로 배치해 주세요.
       top: 50%;
       transform: translateY(-50%);
-      font-size: 2rem;
+      font-size: 1.5rem;
       color: gray;
       cursor: pointer;
       width: auto;
 
       &.left {
-        left: 10px;
+        left: 380px;
       }
 
       &.right {
-        right: 10px;
+        right: 380px;
       }
     }
 
@@ -203,27 +140,4 @@ const BannerBlock = styled.div`
     font-weight: 600;
     line-height: 1.8rem;
   }
-`;
-
-const DotsBlock = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: absolute;
-
-  padding: 10px;
-`;
-
-const Dot = styled.div`
-  width: 10px;
-  height: 10px;
-  border-radius: 50%;
-  background: #ccc;
-  margin: 0 5px;
-
-  ${(props) =>
-    props.active &&
-    css`
-      background: #333;
-    `}
 `;
